moved currentjobid to jobcollection, refactor to comply with change
